args = "var x = new XML(" + args + ");";
When evaluating this line, toString() method of the 'args' object is invoked.

If the 'args' is an instance of an OMElement, then the OMElement is serialized to a String..

This is an unnecessary overhead, since Axiom Objects can be directly passed into the E4X impl. 

This issue is fixed in this commit


git-svn-id: http://svn.wso2.org/repos/wso2/trunk/wsf@1728 a5903396-d722-0410-b921-86c7d4935375